frontend/requirements.txt
streamlit
requests
streamlit-chat==0.1.1
python-dotenv